Converted main.dart to login page & added extra parameters to sign up
